initial commit
discord verification bot.

the bot listens to messages on a specific discord server channel for an email address. When a email address is input the bot calls out to squareAPI to determine if the email is associated with an existing customer. If the email is associated with an existing customer the guild members role is updated to a 'verified' role else the user receives a message response stating the email is not associated with a subscription.

Additionally the bot has a /verify command that allows guild users with an 'admin' role. This command displays a dropdown select menu to select users with an 'unverified' role and update their role to 'verified'

Initial commit